+++ src/main/kotlin/home/dj/domain/request/InvoiceRequest.kt
@@ -2,7 +2,6 @@ package home.dj.domain.request
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import home.dj.domain.CostCategory
-import home.dj.domain.UtilityInvoice
 import io.micronaut.serde.annotation.Serdeable
 import java.time.LocalDate
 import javax.validation.Valid
@@ -22,5 +21,5 @@ data class InvoiceRequest(
     @field:JsonProperty("file_name")
     val fileName: String,
     @field:JsonProperty("agreement_id")
-    val agreementId: Int
+    val agreementId: Long
 )
